#17 : Fixing client receive

+++ server/main.cpp
@@ -67,7 +67,7 @@ int main(int argc, char **argv)
     server_thread->start();
 
     std::cout << "Will wait for " << nb_players << " players to join." << std::endl;
-    while (server->clients.size() < (size_t)nb_players);
+    while ((int)server->clients.size() < nb_players);
     server->refuseAdditionalClients();
     std::cout << "Perfect, " << nb_players << " joined !" << std::endl;
     std::cout << "Initializating game players..." << std::endl;
